Format the core with rustfmt and clear clippy
The core had never been through either, which is unusual enough in a public
Rust repo to look like an oversight. Doing it now, in its own commit, so the
CI change that gates on both is readable.

rustfmt touched 15 files. Most of it is line wrapping, but doh.rs is a real
fix: the whole `mod transport` body was indented one level short.

Six clippy warnings, all genuine:

- dns.rs cast a usize to usize.
- net.rs converted four addresses to themselves. smoltcp 0.13 aliases
  Ipv4Address/Ipv6Address to std's types, so Ipv4Address::from(s) is s.
- DnsCache had len() with no is_empty().

Pins the compiler in core/rust-toolchain.toml, so CI, contributors and release
builds all use 1.97.1 rather than whatever rustup last installed. Cargo.toml
keeps rust-version = 1.77 as the floor the source actually needs.

101 tests still pass; clippy is clean under -D warnings.

+++ core/src/dns.rs
@@ -153,7 +153,7 @@ fn read_name(buf: &[u8], start: usize) -> Option<(String, usize)> {
             if jumps > 16 {
                 return None; // cyclic or absurdly chained compression
             }
-            let ptr = (((len & 0x3f) as usize) << 8) | buf[pos + 1] as usize;
+            let ptr = ((len & 0x3f) << 8) | buf[pos + 1] as usize;
             if end_of_name.is_none() {
                 end_of_name = Some(pos + 2);
             }
